requirements.txt
# Core Dependencies
telethon>=1.34.0
python-dotenv>=1.0.0

# Database
pymongo>=4.6.0

# Timezone Support
pytz>=2024.1
